Remove unnecessary boxing and unboxing

+++ src/test/java/com/avaje/tests/query/sqlquery/SqlQueryTests.java
@@ -139,7 +139,7 @@ public class SqlQueryTests extends BaseTestCase {
       public boolean accept(SqlRow bean) {
         count.incrementAndGet();
         Integer id = bean.getInteger("id");
-        return id.intValue() < 3;
+        return id < 3;
       }
     });
